+++ com.alibaba.testme.dao/src/main/java/com/alibaba/testme/dao/TestunitDao.java
@@ -27,13 +27,13 @@ public interface TestunitDao {
      * @param id
      * @return
      */
-    public int deleteTestunitDO(Long id);
+    public int deleteTestunitDO(Integer id);
 
     /**
      * @param id
      * @return
      */
-    public TestunitDO findById(Long id);
+    public TestunitDO findById(Integer id);
 
     /**
      * @param testunitDO
