+++ com.alibaba.testme.service/src/main/java/com/alibaba/testme/service/TestunitFlowCaseDetailService.java
@@ -27,13 +27,13 @@ public interface TestunitFlowCaseDetailService {
      * @param id
      * @return
      */
-    public int deleteTestunitFlowCaseDetailDO(Long id);
+    public int deleteTestunitFlowCaseDetailDO(Integer id);
 
     /**
      * @param id
      * @return
      */
-    public TestunitFlowCaseDetailDO findById(Long id);
+    public TestunitFlowCaseDetailDO findById(Integer id);
 
     /**
      * @param id
