+++ com.alibaba.testme.service/src/main/java/com/alibaba/testme/service/TestunitFlowCaseService.java
@@ -27,13 +27,13 @@ public interface TestunitFlowCaseService {
      * @param id
      * @return
      */
-    public int deleteTestunitFlowCaseDO(Long id);
+    public int deleteTestunitFlowCaseDO(Integer id);
 
     /**
      * @param id
      * @return
      */
-    public TestunitFlowCaseDO findById(Long id);
+    public TestunitFlowCaseDO findById(Integer id);
 
     /**
      * @param id
